requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
click==8.0.1
Django==3.2.5
django-js-asset==1.2.2
django-mptt==0.12.0
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.3.1
psycopg2==2.9.1
pytz==2021.1
regex==2021.7.6
sqlparse==0.4.1
tomli==1.1.0
